fix remain type

+++ ServiceAbonents/Dtos/SwitchTarifDto.cs
@@ -4,9 +4,9 @@ namespace ServiceAbonents.Dtos
 {
     public class SwitchTarifDto
     {
-        public short? Minutes { get; set; }
-        public short? Gigabytes { get; set; }
-        public short? Sms { get; set; }
+        public decimal? Minutes { get; set; }
+        public decimal? Gigabytes { get; set; }
+        public decimal? Sms { get; set; }
         public bool? UnlimVideo { get; set; }
         public bool? UnlimSocials { get; set; }
         public bool? UnlimMusic { get; set; }
